requirements.txt
requests
beautifulsoup4
pandas
tqdm
selenium
argparse
streamlit
python-dotenv
spacy
tqdm